+++ CHANGELOG.md
@@ -1,3 +1,4 @@
 # 1.0.2
 + Module Items now have proper inventory items
 + Fixed having the wrong pickup items for the modules
++ Changed the type clarification from the tooltip to the name